+++ languagetool-wikipedia/src/main/java/org/languagetool/dev/dumpcheck/PlainTextSentenceSource.java
@@ -38,7 +38,7 @@ public class PlainTextSentenceSource extends SentenceSource {
   private final Scanner scanner;
 
   // Each sentence is one article, but count anyway so it's coherent with what the Wikipedia code does:
-  private int articleCount = 0;
+  private Long articleCount = 0L;
   private String currentUrl = null;
 
   public PlainTextSentenceSource(InputStream textInput, Language language) {
